+++ src/main/java/com/habitproject/service/habit/HabitService.java
@@ -1,7 +1,7 @@
-package com.habitproject.service;
+package com.habitproject.service.habit;
 
-import com.habitproject.web.EntityStatusReturns.HabitListStatusReturn;
-import com.habitproject.web.EntityStatusReturns.HabitStatusReturn;
+import com.habitproject.web.habit.HabitListStatusReturn;
+import com.habitproject.web.habit.HabitStatusReturn;
 import com.habitproject.web.habit.HabitRequestModel;
 import org.springframework.http.HttpStatus;
