+++ src/main/java/com/habitproject/service/habit/HabitServiceImpl.java
@@ -1,9 +1,9 @@
-package com.habitproject.service;
+package com.habitproject.service.habit;
 
 import com.habitproject.persistence.habit.HabitEntity;
 import com.habitproject.persistence.habit.HabitRepository;
-import com.habitproject.web.EntityStatusReturns.HabitListStatusReturn;
-import com.habitproject.web.EntityStatusReturns.HabitStatusReturn;
+import com.habitproject.web.habit.HabitListStatusReturn;
+import com.habitproject.web.habit.HabitStatusReturn;
 import com.habitproject.web.habit.HabitRequestModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
